// packages/client/src/constants/routes.ts
export enum Routes {
  Homepage = '/',
  ProfileRoot = '/profile',
  Profile = '/profile/:tab?/:subTab?',
  Casino = '/casino',
  Games = '/games',
  GamesSlug = '/game/:id',
  InPlays = '/inplays',
  Sports = '/sports',
  LiveCasino = '/live-casino',
  ESports = '/e-sports',
  Promotions = '/promotions',
  PromotionSlug = '/promotions/:slug',
}